Refactor voxelization.py into a voxelization/ package
Split the single voxelization module into a package so a future subgrid
backend can sit beside the marching-squares one without duplicating the
shared coordinate helpers or orchestration. Pure structural move: function
bodies are lifted verbatim, no behavior or public API change.

- _coords.py: grid coordinate helpers + CenteringMode
- marching_squares.py: volume-fraction discretization algorithm
- sampling.py: stochastic occupancy realization (purves/beta)
- tree.py: VoxelizedTree + voxelize_tree orchestration
- __init__.py: thin re-export of the public API

trees.py is unchanged (public names still import from
fastfuels_core.voxelization); test imports of internal helpers now point
at their submodules.

Closes #77
